refactor: 💡 VideoOverlayToggle 분리
